requirements.txt
Flask==3.0.3
twelvelabs
requests==2.31.0
python-dotenv==1.0.1
gunicorn